# internal/adapters/http/in/transaction_validation_handler.go
// Copyright (c) 2026 Lerian Studio. All rights reserved.
// Use of this source code is governed by the Elastic License 2.0
// that can be found in the LICENSE file.

package in

//go:generate mockgen -source=transaction_validation_handler.go -destination=mocks/transaction_validation_service_mock.go -package=mocks

import (
	"context"
	"errors"
	"fmt"
	"time"

	libCommons "github.com/LerianStudio/lib-commons/v2/commons"
	libHTTP "github.com/LerianStudio/lib-commons/v2/commons/net/http"
	libOpentelemetry "github.com/LerianStudio/lib-commons/v2/commons/opentelemetry"
	"github.com/gofiber/fiber/v2"
	"github.com/google/uuid"
	"go.opentelemetry.io/otel/trace"

	"tracer/internal/services/query"
	"tracer/pkg/constant"
	"tracer/pkg/logging"
	"tracer/pkg/model"
	pkgHTTP "tracer/pkg/net/http"
	"tracer/pkg/validation"
)

// TransactionValidationService defines the interface for transaction validation operations.
// Interface defined locally per Ring pattern.
// NOTE: Transaction validations are immutable per SOX/GLBA requirements - only read operations.
type TransactionValidationService interface {
	GetTransactionValidation(ctx context.Context, id uuid.UUID) (*model.TransactionValidation, error)
	ListTransactionValidations(ctx context.Context, filters *model.TransactionValidationFilters) (*query.ListTransactionValidationsResult, error)
}

// TransactionValidationHandler handles HTTP requests for transaction validation operations.
type TransactionValidationHandler struct {
	service TransactionValidationService
}

// NewTransactionValidationHandler creates a new transaction validation handler.
func NewTransactionValidationHandler(service TransactionValidationService) *TransactionValidationHandler {
	return &TransactionValidationHandler{
		service: service,
	}
}

// GetTransactionValidation godoc
//
//	@Summary		Get a transaction validation record by ID
//	@Description	Retrieves a transaction validation record by its unique identifier.
//	@ID				getValidation
//	@Tags			validations
//	@Accept			json
//	@Produce		json
//	@Security		ApiKeyAuth
//	@Param			id			path		string					true	"Transaction Validation ID (UUID)"	Format(uuid)
//	@Success		200			{object}	model.TransactionValidation	"Transaction validation retrieved successfully"
//	@Failure		400			{object}	api.ErrorResponse		"Invalid transaction validation ID"
//	@Failure		401			{object}	api.ErrorResponse		"Unauthorized"
//	@Failure		404			{object}	api.ErrorResponse		"Transaction validation not found"
//	@Failure		500			{object}	api.ErrorResponse		"Internal server error"
//	@Router			/v1/validations/{id} [get]
func (h *TransactionValidationHandler) GetTransactionValidation(c *fiber.Ctx) error {
	ctx := c.UserContext()
	logger, tracer, _, _ := libCommons.NewTrackingFromContext(ctx)

	ctx, span := tracer.Start(ctx, "handler.transaction-validation.get")
	defer span.End()

	logger = logging.WithTrace(ctx, logger)

	// Parse transaction validation ID from path
	idParam := c.Params("id")

	id, err := uuid.Parse(idParam)
	if err != nil {
		libOpentelemetry.HandleSpanBusinessErrorEvent(&span, "Invalid transaction validation ID", err)

		return pkgHTTP.BadRequestWithMessage(c, "TRC-0007", "Invalid Path Parameter", "Invalid transaction validation ID format")
	}

	logger.WithFields(
		"operation", "handler.transaction-validation.get",
		"transaction_validation.id", id.String(),
	).Info("Getting transaction validation record")

	result, err := h.service.GetTransactionValidation(ctx, id)
	if err != nil {
		return handleTransactionValidationServiceError(c, &span, err)
	}

	logger.WithFields(
		"operation", "handler.transaction-validation.get",
		"transaction_validation.id", result.ID.String(),
		"transaction_validation.decision", string(result.Decision),
	).Info("Transaction validation record retrieved")

	return libHTTP.OK(c, result)
}

// ListTransactionValidations godoc
//
//	@Summary		List transaction validation records
//	@Description	Lists transaction validation records with cursor-based pagination and filters.
//	@ID				listValidations
//	@Tags			validations
//	@Accept			json
//	@Produce		json
//	@Security		ApiKeyAuth
//	@Param			limit			query		int								false	"Max items per page (1-1000, default: 100)"		minimum(1)	maximum(1000)
//	@Param			cursor			query		string							false	"Pagination cursor from previous response"
//	@Param			sortBy			query		string							false	"Field to sort by (default: createdAt)"		Enums(createdAt, processingTimeMs)
//	@Param			sortOrder		query		string							false	"Sort direction (default: DESC)"				Enums(ASC, DESC)
//	@Param			startDate		query		string							false	"Filter from this date (RFC3339)"				Format(date-time)
//	@Param			endDate			query		string							false	"Filter to this date (RFC3339)"					Format(date-time)
//	@Param			decision		query		string							false	"Filter by decision"							Enums(ALLOW, DENY, REVIEW)
//	@Param			accountId		query		string							false	"Filter by account ID (UUID)"					Format(uuid)
//	@Param			matchedRuleId	query		string							false	"Filter by matched rule ID (UUID)"				Format(uuid)
//	@Param			exceededLimitId	query		string							false	"Filter by exceeded limit ID (UUID)"			Format(uuid)
//	@Param			segmentId		query		string							false	"Filter by segment ID (UUID)"					Format(uuid)
//	@Param			portfolioId		query		string							false	"Filter by portfolio ID (UUID)"					Format(uuid)
//	@Param			transactionType	query		string							false	"Filter by transaction type"					Enums(CARD, WIRE, PIX, CRYPTO)
//	@Success		200				{object}	ListTransactionValidationsResponse	"Transaction validations listed successfully"
//	@Failure		400				{object}	api.ErrorResponse				"Invalid parameters"
//	@Failure		401				{object}	api.ErrorResponse				"Unauthorized"
//	@Failure		500				{object}	api.ErrorResponse				"Internal server error"
//	@Router			/v1/validations [get]
func (h *TransactionValidationHandler) ListTransactionValidations(c *fiber.Ctx) error {
	ctx := c.UserContext()
	logger, tracer, _, _ := libCommons.NewTrackingFromContext(ctx)

	ctx, span := tracer.Start(ctx, "handler.transaction-validation.list")
	defer span.End()

	logger = logging.WithTrace(ctx, logger)

	// Parse query parameters into input struct
	var input ListTransactionValidationsInput

	if err := c.QueryParser(&input); err != nil {
		libOpentelemetry.HandleSpanError(&span, "Failed to parse query parameters", err)

		return pkgHTTP.BadRequestWithMessage(c, "TRC-0006", "Invalid Query Parameter", "Invalid query parameters")
	}

	// Validate before applying defaults to ensure fail-fast behavior
	if err := input.Validate(); err != nil {
		libOpentelemetry.HandleSpanBusinessErrorEvent(&span, "Validation failed", err)

		// Check if error is ValidationError with specific code
		var validationErr *ValidationError
		if errors.As(err, &validationErr) {
			return pkgHTTP.BadRequestWithMessage(c, validationErr.Code, "Invalid Transaction Validation Filters", validationErr.Message)
		}

		return pkgHTTP.BadRequestWithMessage(c, "TRC-0250", "Invalid Transaction Validation Filters", err.Error())
	}

	// Apply defaults after validation
	input.SetDefaults()

	logger.WithFields(
		"operation", "handler.transaction-validation.list",
		"list.limit", input.Limit,
		"list.cursor", input.Cursor,
		"list.sort_by", input.SortBy,
		"list.sort_order", input.SortOrder,
	).Info("Listing transaction validation records")

	// Convert to service filter
	filters, err := ToTransactionValidationFilters(&input)
	if err != nil {
		libOpentelemetry.HandleSpanBusinessErrorEvent(&span, "Invalid filters", err)

		return pkgHTTP.BadRequestWithMessage(c, "TRC-0250", "Invalid Transaction Validation Filters", err.Error())
	}

	result, err := h.service.ListTransactionValidations(ctx, filters)
	if err != nil {
		return handleTransactionValidationServiceError(c, &span, err)
	}

	// Convert to response
	response := ToListTransactionValidationsResponse(result)

	logger.WithFields(
		"operation", "handler.transaction-validation.list",
		"list.count", len(response.TransactionValidations),
		"list.has_more", response.HasMore,
	).Info("Transaction validation records listed")

	return libHTTP.OK(c, response)
}

// handleTransactionValidationServiceError converts service errors to appropriate HTTP responses.
func handleTransactionValidationServiceError(c *fiber.Ctx, span *trace.Span, err error) error {
	switch {
	case errors.Is(err, constant.ErrTransactionValidationNotFound):
		libOpentelemetry.HandleSpanBusinessErrorEvent(span, "Transaction validation not found", err)

		return libHTTP.NotFound(c, "TRC-0251", "Not Found", "Transaction validation not found")
	case errors.Is(err, constant.ErrInvalidTransactionValidationFilters):
		libOpentelemetry.HandleSpanBusinessErrorEvent(span, "Invalid transaction validation filters", err)

		return pkgHTTP.BadRequestWithMessage(c, "TRC-0250", "Invalid Transaction Validation Filters", "Invalid transaction validation filters")
	case errors.Is(err, constant.ErrInvalidCursor):
		libOpentelemetry.HandleSpanBusinessErrorEvent(span, "Invalid cursor", err)

		return libHTTP.BadRequest(c, fiber.Map{
			"code":    "TRC-0044",
			"title":   "Bad Request",
			"message": "Invalid pagination cursor",
		})
	default:
		libOpentelemetry.HandleSpanError(span, "Operation failed", err)

		return libHTTP.InternalServerError(c, "TRC-0004", "Internal Server Error", "An unexpected error occurred")
	}
}

// ListTransactionValidationsInput represents query parameters for listing transaction validation records.
// Uses cursor-based pagination for consistent, efficient pagination with large datasets.
type ListTransactionValidationsInput struct {
	Limit           *int   `query:"limit"`
	Cursor          string `query:"cursor"`
	SortBy          string `query:"sortBy"`
	SortOrder       string `query:"sortOrder"`
	StartDate       string `query:"startDate"`
	EndDate         string `query:"endDate"`
	Decision        string `query:"decision"`
	AccountID       string `query:"accountId"`
	MatchedRuleID   string `query:"matchedRuleId"`
	ExceededLimitID string `query:"exceededLimitId"`
	SegmentID       string `query:"segmentId"`
	PortfolioID     string `query:"portfolioId"`
	TransactionType string `query:"transactionType"`
}

// validateUUID checks if a string is a valid UUID format.
// Returns nil if value is empty (optional field) or valid UUID.
func validateUUID(value, fieldName string) error {
	if value == "" {
		return nil
	}

	if _, err := uuid.Parse(value); err != nil {
		return fmt.Errorf("%s must be a valid UUID", fieldName)
	}

	return nil
}

// SetDefaults sets default values for optional fields after validation.
// Note: Only applies sort defaults when cursor is not present, as cursor already contains sort configuration.
func (i *ListTransactionValidationsInput) SetDefaults() {
	if i.Limit == nil {
		defaultLimit := model.DefaultTransactionValidationFilterLimit
		i.Limit = &defaultLimit
	}

	// Only apply sort defaults when not using cursor pagination
	// Cursor already contains sort configuration from the original request (TRC-0045)
	if i.Cursor == "" {
		if i.SortBy == "" {
			i.SortBy = "createdAt"
		}

		// Normalize sortOrder to uppercase and apply default if empty
		i.SortOrder = NormalizeSortOrder(i.SortOrder, "DESC")
	}
}

// Validate checks if the input is valid.
// Validates before defaults are applied to ensure fail-fast behavior.
func (i *ListTransactionValidationsInput) Validate() error {
	// Validate pagination limit (TRC-0040, TRC-0041)
	if err := ValidatePaginationLimit(i.Limit, model.MaxTransactionValidationFilterLimit); err != nil {
		return err
	}

	// Validate cursor consistency (TRC-0045)
	if err := ValidateCursorConsistency(i.Cursor, i.SortBy, i.SortOrder); err != nil {
		return err
	}

	// Validate sortBy whitelist (TRC-0043)
	allowedSortFields := []string{"createdAt", "processingTimeMs"}
	if err := ValidateSortBy(i.SortBy, allowedSortFields); err != nil {
		return err
	}

	// Validate sortOrder enum (TRC-0042)
	if err := ValidateSortOrder(i.SortOrder); err != nil {
		return err
	}

	if err := i.validateDecision(); err != nil {
		return err
	}

	if err := i.validateDates(); err != nil {
		return err
	}

	if err := i.validateUUIDFilters(); err != nil {
		return err
	}

	return i.validateTransactionType()
}

func (i *ListTransactionValidationsInput) validateDecision() error {
	if i.Decision != "" && !model.Decision(i.Decision).IsValid() {
		return errors.New("decision must be one of [ALLOW, DENY, REVIEW]")
	}

	return nil
}

func (i *ListTransactionValidationsInput) validateDates() error {
	startDate, err := validation.ParseRFC3339Timestamp(i.StartDate, "startDate")
	if err != nil {
		return &ValidationError{
			Code:    "TRC-0020",
			Message: err.Error(),
		}
	}

	endDate, err := validation.ParseRFC3339Timestamp(i.EndDate, "endDate")
	if err != nil {
		return &ValidationError{
			Code:    "TRC-0020",
			Message: err.Error(),
		}
	}

	// Use user-friendly error message for date range validation
	if !startDate.IsZero() && !endDate.IsZero() && startDate.After(endDate) {
		return &ValidationError{
			Code:    "TRC-0023",
			Message: "startDate must be before or equal to endDate",
		}
	}

	return nil
}

func (i *ListTransactionValidationsInput) validateUUIDFilters() error {
	if err := validateUUID(i.AccountID, "accountId"); err != nil {
		return err
	}

	if err := validateUUID(i.MatchedRuleID, "matchedRuleId"); err != nil {
		return err
	}

	if err := validateUUID(i.ExceededLimitID, "exceededLimitId"); err != nil {
		return err
	}

	if err := validateUUID(i.SegmentID, "segmentId"); err != nil {
		return err
	}

	return validateUUID(i.PortfolioID, "portfolioId")
}

func (i *ListTransactionValidationsInput) validateTransactionType() error {
	if i.TransactionType != "" && !model.TransactionType(i.TransactionType).IsValid() {
		return errors.New("transactionType must be one of [CARD, WIRE, PIX, CRYPTO]")
	}

	return nil
}

// ToTransactionValidationFilters converts the HTTP input to service filters.
func ToTransactionValidationFilters(input *ListTransactionValidationsInput) (*model.TransactionValidationFilters, error) {
	limit := model.DefaultTransactionValidationFilterLimit
	if input.Limit != nil {
		limit = *input.Limit
	}

	filters := &model.TransactionValidationFilters{
		Limit:     limit,
		Cursor:    input.Cursor,
		SortBy:    input.SortBy,
		SortOrder: input.SortOrder,
	}

	// Parse start date
	if input.StartDate != "" {
		startDate, err := time.Parse(time.RFC3339, input.StartDate)
		if err != nil {
			return nil, err
		}

		filters.StartDate = startDate
	}

	// Parse end date
	if input.EndDate != "" {
		endDate, err := time.Parse(time.RFC3339, input.EndDate)
		if err != nil {
			return nil, err
		}

		filters.EndDate = endDate
	}

	// Parse decision
	if input.Decision != "" {
		decision := model.Decision(input.Decision)
		filters.Decision = &decision
	}

	// Parse account ID
	if input.AccountID != "" {
		accountID, err := uuid.Parse(input.AccountID)
		if err != nil {
			return nil, err
		}

		filters.AccountID = &accountID
	}

	// Parse matched rule ID
	if input.MatchedRuleID != "" {
		matchedRuleID, err := uuid.Parse(input.MatchedRuleID)
		if err != nil {
			return nil, err
		}

		filters.MatchedRuleID = &matchedRuleID
	}

	// Parse exceeded limit ID
	if input.ExceededLimitID != "" {
		exceededLimitID, err := uuid.Parse(input.ExceededLimitID)
		if err != nil {
			return nil, err
		}

		filters.ExceededLimitID = &exceededLimitID
	}

	// Parse segment ID
	if input.SegmentID != "" {
		segmentID, err := uuid.Parse(input.SegmentID)
		if err != nil {
			return nil, err
		}

		filters.SegmentID = &segmentID
	}

	// Parse portfolio ID
	if input.PortfolioID != "" {
		portfolioID, err := uuid.Parse(input.PortfolioID)
		if err != nil {
			return nil, err
		}

		filters.PortfolioID = &portfolioID
	}

	// Parse transaction type
	if input.TransactionType != "" {
		transactionType := model.TransactionType(input.TransactionType)
		filters.TransactionType = &transactionType
	}

	return filters, nil
}

// ValidationSummary represents a summary of a validation record in list responses.
// Per API Design v1.3.2: List endpoint returns ValidationSummary, not full TransactionValidation.
// Fields are flattened (accountId at root, not nested in account object).
type ValidationSummary struct {
	ID               uuid.UUID             `json:"validationId" swaggertype:"string" format:"uuid"`
	Decision         model.Decision        `json:"decision"`
	Reason           string                `json:"reason"`
	Amount           int64                 `json:"amount"`
	Currency         string                `json:"currency"`
	TransactionType  model.TransactionType `json:"transactionType"`
	AccountID        uuid.UUID             `json:"accountId" swaggertype:"string" format:"uuid"`
	SegmentID        *uuid.UUID            `json:"segmentId,omitempty" swaggertype:"string" format:"uuid"`
	PortfolioID      *uuid.UUID            `json:"portfolioId,omitempty" swaggertype:"string" format:"uuid"`
	MatchedRuleIDs   []uuid.UUID           `json:"matchedRuleIds" swaggertype:"array,string" format:"uuid"`
	ExceededLimitIDs []uuid.UUID           `json:"exceededLimitIds" swaggertype:"array,string" format:"uuid"`
	ProcessingTimeMs int64                 `json:"processingTimeMs"`
	CreatedAt        string                `json:"createdAt" format:"date-time"`
}

// ToValidationSummary converts a TransactionValidation to a ValidationSummary.
func ToValidationSummary(tv *model.TransactionValidation) *ValidationSummary {
	if tv == nil {
		return nil
	}

	summary := &ValidationSummary{
		ID:               tv.ID,
		Decision:         tv.Decision,
		Reason:           tv.Reason,
		Amount:           tv.Amount,
		Currency:         tv.Currency,
		TransactionType:  tv.TransactionType,
		AccountID:        tv.Account.ID,
		MatchedRuleIDs:   ensureUUIDSlice(tv.MatchedRuleIDs),
		ProcessingTimeMs: tv.ProcessingTimeMs,
		CreatedAt:        tv.CreatedAt.Format(time.RFC3339),
	}

	// Extract exceeded limit IDs from limit usage details
	exceededLimitIDs := make([]uuid.UUID, 0)

	for _, detail := range tv.LimitUsageDetails {
		if detail.Exceeded {
			exceededLimitIDs = append(exceededLimitIDs, detail.LimitID)
		}
	}

	summary.ExceededLimitIDs = exceededLimitIDs

	// Extract optional segment and portfolio IDs
	if tv.Segment != nil {
		summary.SegmentID = &tv.Segment.ID
	}

	if tv.Portfolio != nil {
		summary.PortfolioID = &tv.Portfolio.ID
	}

	return summary
}

// ListTransactionValidationsResponse represents the response for listing transaction validation records.
// Uses cursor-based pagination with nextCursor and hasMore fields.
// Per API Design v1.3.2: Returns ValidationSummary objects with flattened fields.
type ListTransactionValidationsResponse struct {
	TransactionValidations []*ValidationSummary `json:"transactionValidations"`
	NextCursor             string               `json:"nextCursor,omitempty"`
	HasMore                bool                 `json:"hasMore"`
}

// ToListTransactionValidationsResponse converts the service result to HTTP response.
func ToListTransactionValidationsResponse(result *query.ListTransactionValidationsResult) *ListTransactionValidationsResponse {
	summaries := make([]*ValidationSummary, len(result.TransactionValidations))
	for i, tv := range result.TransactionValidations {
		summaries[i] = ToValidationSummary(tv)
	}

	return &ListTransactionValidationsResponse{
		TransactionValidations: summaries,
		NextCursor:             result.NextCursor,
		HasMore:                result.HasMore,
	}
}

// ensureUUIDSlice returns an empty slice if input is nil, ensuring JSON serializes as [] not null.
func ensureUUIDSlice(ids []uuid.UUID) []uuid.UUID {
	if ids == nil {
		return []uuid.UUID{}
	}

	return ids
}